jenkins: support token authentication
Allow a username and token to be specified in the Jenkins configuration.

To enable this, add a few helper methods for JenkinsBackend that spit out
Authorization headers and wrap the hyper client's get() and post() methods
to use those headers.

Also, update the example configuration, 'openpower.toml', to include fake
credentials.

Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>