Allow specifying a name for each job
Currently we just report the name of the job as the name of the check we
report to Patchwork.  This is problematic if we want to run the same job
multiple times with different configurations (Patchwork will just pick
one), so allow specifying an explicit name for each job we kick off.

Signed-off-by: Russell Currey <ruscur@russell.cc>