jenkins: Rename "build_url" to "build_handle"
In preparation for genericising our CI backend API, change references to
"build_url" to the more generic "build_handle", because the fact that it's
a URL is really a Jenkins-specific detail. Snowpatch cares about the fact
that a results URL is a URL, but not so much the build URL, which could be
any arbitrary identifier really.

Also some random formatting cleanups.

Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>